Add proper dynamic angle bracket invocation detection.
Prior to this we assumed any element node that started with a capital
letter was a component invocation, but that left out a significant
number of scenarios (e.g. yielded contextual components).

This brings a block param scope tracking system so that we can be
certain that dynamic angle invocations are detected correctly.