Ensure simple property paths are not mutated.
Prior to this, the AST transform would add a named argument to _every_
mustache statement. This meant that simple/normal property lookups were
getting rewritten from `{{foo}}` to `{{foo
debugTemplateInvocationSite=(hash)}}`. When Ember sees a mustache
statement that has arguments (positional or named) it **always** assumes
it is a component invocation. Which means that when we rewrite `{{foo}}`
from above, we have completely broken the ability to render any dynamic
value in the template 😫.

The changes in this commit tweak things so that when we are dealing with
an ambiguous mustache invocation (we can't tell if its a property
lookup, a helper invocation, a component invocation, etc) that isn't a
built-in keyword we use a runtime helper to determine if the path refers
to a component or helper. This allows us to handle a simple mustache
like `{{foo}}` by rewritting to something like (whitespace added for
clarity, but the result will not include additional whitespace):

```hbs
{{#if (-template-invocation-info-is-component-or-helper  foo)}}
  {{foo debugTemplateInvocationSite=(hash ...)}}
{{else}}
  {{foo}}
{{/if}}
```